Codereview move function extract_vuln_packages_from_intel to utils
