+++ src/vuln_analysis/functions/cve_verify_vuln_package.py
@@ -42,8 +42,8 @@ from exploit_iq_commons.data_models.dependencies import (
 from exploit_iq_commons.utils.git_utils import get_repo_path_with_ref
 from exploit_iq_commons.logging.loggers_factory import LoggingFactory, trace_id
 from nat.builder.context import Context
-from vuln_analysis.utils.package_identifier import _RPM_NEVRA_RE
 from exploit_iq_commons.utils.dep_tree import Ecosystem
+from vuln_analysis.utils.intel_utils import extract_vuln_packages_from_intel
 
 logger = LoggingFactory.get_agent_logger(__name__)
 
@@ -371,127 +371,6 @@ class CVEVerifyVulnPackageConfig(FunctionBaseConfig, name="cve_verify_vuln_packa
         default="checklist_llm",
         description="Name of the LLM to use for version comparison"
     )
-
-
-def extract_vuln_packages_from_intel(cve_intel: CveIntel) -> list[dict[str, Any]]:
-    """
-    Extract vulnerable package names and version ranges from intel sources.
-
-    Args:
-        cve_intel: CveIntel object containing NVD, GHSA, and RHSA data
-
-    Returns:
-        List of dicts with package info:
-        {
-            "source": "nvd" | "ghsa" | "rhsa",
-            "package": str,
-            "ecosystem": str | None,
-            "version_start_incl": str | None,
-            "version_start_excl": str | None,
-            "version_end_incl": str | None,
-            "version_end_excl": str | None,
-            "vulnerable_range": str | None,  # GHSA format like "< 2.9.2"
-            "first_patched": str | None,
-            "fix_state": str | None,  # RHSA fix_state ("affected", "not affected", etc.)
-        }
-    """
-    packages: list[dict[str, Any]] = []
-
-    # Extract from NVD configurations
-    if cve_intel.nvd and cve_intel.nvd.configurations:
-        for config in cve_intel.nvd.configurations:
-            packages.append({
-                "source": "nvd",
-                "package": config.package,
-                "ecosystem": config.system,
-                "version_start_incl": config.versionStartIncluding,
-                "version_start_excl": config.versionStartExcluding,
-                "version_end_incl": config.versionEndIncluding,
-                "version_end_excl": config.versionEndExcluding,
-                "vulnerable_range": None,
-                "first_patched": None,
-                "fix_state": None,
-            })
-
-    # Extract from GHSA vulnerabilities
-    if cve_intel.ghsa and cve_intel.ghsa.vulnerabilities:
-        for vuln in cve_intel.ghsa.vulnerabilities:
-            # Handle both dict and object formats
-            if isinstance(vuln, dict):
-                pkg_info = vuln.get("package", {})
-                pkg_name = pkg_info.get("name") if isinstance(pkg_info, dict) else None
-                pkg_ecosystem = pkg_info.get("ecosystem") if isinstance(pkg_info, dict) else None
-                vuln_range = vuln.get("vulnerable_version_range")
-                first_patched = vuln.get("first_patched_version")
-            else:
-                pkg_info = getattr(vuln, "package", {})
-                pkg_name = pkg_info.get("name") if isinstance(pkg_info, dict) else None
-                pkg_ecosystem = pkg_info.get("ecosystem") if isinstance(pkg_info, dict) else None
-                vuln_range = getattr(vuln, "vulnerable_version_range", None)
-                first_patched = getattr(vuln, "first_patched_version", None)
-
-            if pkg_name:
-                packages.append({
-                    "source": "ghsa",
-                    "package": pkg_name,
-                    "ecosystem": pkg_ecosystem,
-                    "version_start_incl": None,
-                    "version_start_excl": None,
-                    "version_end_incl": None,
-                    "version_end_excl": None,
-                    "vulnerable_range": vuln_range,
-                    "first_patched": first_patched,
-                    "fix_state": None,
-                })
-
-    # Extract from RHSA package_state (affected packages with fix status)
-    if cve_intel.rhsa and cve_intel.rhsa.package_state:
-        for ps in cve_intel.rhsa.package_state:
-            if ps.package_name:
-                packages.append({
-                    "source": "rhsa",
-                    "package": ps.package_name,
-                    "ecosystem": "rpm",
-                    "version_start_incl": None,
-                    "version_start_excl": None,
-                    "version_end_incl": None,
-                    "version_end_excl": None,
-                    "vulnerable_range": None,
-                    "first_patched": None,
-                    "fix_state": ps.fix_state,
-                })
-
-    # Extract from RHSA affected_release (fixed versions)
-    if cve_intel.rhsa and hasattr(cve_intel.rhsa, 'affected_release') and cve_intel.rhsa.affected_release:
-        for entry in cve_intel.rhsa.affected_release:
-            pkg_str = entry.get("package") if isinstance(entry, dict) else getattr(entry, "package", None)
-            if pkg_str:
-                # Parse NEVRA format: name-[epoch:]version-release.arch
-                m = _RPM_NEVRA_RE.match(pkg_str)
-                if m:
-                    name = m.group(1)
-                    version = m.group(3)
-                    release_arch = m.group(4)
-                    release = release_arch.rsplit(".", 1)[0] if "." in release_arch else release_arch
-                    ver_rel = f"{version}-{release}"
-                else:
-                    name = pkg_str
-                    ver_rel = None
-
-                packages.append({
-                    "source": "rhsa",
-                    "package": name,
-                    "ecosystem": "rpm",
-                    "version_start_incl": None,
-                    "version_start_excl": None,
-                    "version_end_incl": None,
-                    "version_end_excl": ver_rel,  # Fixed version = end of vulnerable range
-                    "vulnerable_range": f"< {ver_rel}" if ver_rel else None,
-                    "first_patched": ver_rel,
-                    "fix_state": None,
-                })
-
-    return packages
 
 
 @register_function(config_type=CVEVerifyVulnPackageConfig, framework_wrappers=[LLMFrameworkEnum.LANGCHAIN])
